Converting Environment from module to testsuite.
Now Environment is a special testsuite, which startup and testcases
are executed prior to the execution of other testsuites. Cleanup
is executed after at the very end - can be used to specify post
testing actions.

Known bug: logging is implemented as singleton, which means last
testsuite`s Logger will be used in Environment.cleanup and every
message will be logged at the end of the last testsuite.