Bad Mutex on get_unused_name in DirectoryServer.
Since the mutex was split into two parts - one covering get_unused_name
and the other DirectoryServer.setup itself, it was possible to interleave
getting the names and setting up instances so that the second thread would
fetch unused instance name before the first one is set up.

Setting up global variable storing list of used names and ports would not
work, as it would be visible only from one process, but not others that
were forked. Therefore, going back to one big mutex around get_instance.
Using DirectoryServer.setup is not recommended, as it is not safe.